requirements.txt
Flask>=2.0.0
Flask_Cors
pandas
openai
python-dotenv
tqdm
PyMuPDF
psycopg2-binary
selenium
webdriver-manager
pdf2image
pytesseract
selenium-wire
beautifulsoup4
celery
redis
../classify_presentations
../train_mix_tagger
joblib